Mino bug fix on sql files

+++ dbmodel/ud/ud_gw_trg_edit_man_node.sql
@@ -42,7 +42,7 @@ BEGIN
 
         -- Node ID
         IF (NEW.node_id IS NULL) THEN
-            SELECT max(node_id) INTO node_id_seq FROM node WHERE node_id ~ '^\d+$';
+            SELECT max(node_id::integer) INTO node_id_seq FROM node WHERE node_id ~ '^\d+$';
             PERFORM setval('node_id_seq',node_id_seq,true);
             NEW.node_id:= (SELECT nextval('node_id_seq'));
         END IF;
