+++ dbmodel/ws/ws_gw_trg_edit_man_arc.sql
@@ -22,7 +22,7 @@ BEGIN
     
         -- Arc ID
         IF (NEW.arc_id IS NULL) THEN
-            SELECT max(arc_id) INTO arc_id_seq FROM arc WHERE arc_id ~ '^\d+$';
+            SELECT max(arc_id::integer) INTO arc_id_seq FROM arc WHERE arc_id ~ '^\d+$';
             PERFORM setval('arc_id_seq',arc_id_seq,true);
             NEW.arc_id:= (SELECT nextval('arc_id_seq'));
         END IF;
